dganalytics/connectors/gpc/realtime/functions/queue_stats_5sec/requirements.txt
azure-functions
requests
pandas
msrestazure
azure-mgmt-resource
azure-keyvault
azure-identity